Move objectives to objectives.ts
