+++ src/app/inventory/store/d2-item-factory.service.ts
@@ -4,13 +4,10 @@ import {
   DestinyItemComponent,
   DestinyItemComponentSetOfint64,
   DestinyItemInstanceComponent,
-  DestinyItemObjectivesComponent,
   DestinyItemTalentGridComponent,
-  DestinyObjectiveDefinition,
   DestinyTalentGridDefinition,
   ItemLocation,
   TransferStatuses,
-  DestinyUnlockValueUIStyle,
   DestinyAmmunitionType,
   ItemState,
   DestinyCollectibleComponent,
@@ -24,14 +21,7 @@ import { D2ManifestService } from '../../manifest/manifest-service-json';
 import { NewItemsService } from './new-items.service';
 import { ItemInfoSource } from '../dim-item-info';
 import { t } from 'app/i18next-t';
-import {
-  D2Item,
-  DimPerk,
-  DimObjective,
-  DimFlavorObjective,
-  DimTalentGrid,
-  DimGridNode
-} from '../item-types';
+import { D2Item, DimPerk, DimTalentGrid, DimGridNode } from '../item-types';
 import { D2Store } from '../store-types';
 import { InventoryBuckets } from '../inventory-buckets';
 import { D2StoresService } from '../d2-stores.service';
@@ -42,11 +32,10 @@ import D2SeasonToSource from 'data/d2/seasonToSource.json';
 import D2Events from 'data/d2/events.json';
 import idx from 'idx';
 import { compareBy, chainComparator } from 'app/comparators';
-import memoizeOne from 'memoize-one';
-import { settings } from 'app/settings/settings';
 import { buildStats } from './stats';
 import { buildSockets } from './sockets';
 import { buildMasterwork } from './masterwork';
+import { buildObjectives, buildFlavorObjective } from './objectives';
 
 // Maps tierType to tierTypeName in English
 const tiers = ['Unknown', 'Currency', 'Common', 'Uncommon', 'Rare', 'Legendary', 'Exotic'];
@@ -410,7 +399,7 @@ export function makeItem(
       createdItem.objectives = buildObjectives(
         item,
         objectiveData,
-        defs.Objective,
+        defs,
         uninstancedItemObjectives
       );
     }
@@ -421,7 +410,7 @@ export function makeItem(
 
   try {
     if (objectiveData) {
-      createdItem.flavorObjective = buildFlavorObjective(item, objectiveData, defs.Objective);
+      createdItem.flavorObjective = buildFlavorObjective(item, objectiveData, defs);
     }
   } catch (e) {
     console.error(`Error building flavor objectives for ${createdItem.name}`, item, itemDef, e);
@@ -538,115 +527,6 @@ function getSeason(item: D2Item): number {
   }
 
   return D2Seasons[item.hash] || D2CalculatedSeason || D2CurrentSeason;
-}
-
-const formatterSelector = memoizeOne((language) => new Intl.NumberFormat(language));
-
-function buildObjectives(
-  item: DestinyItemComponent,
-  objectivesMap: { [key: string]: DestinyItemObjectivesComponent },
-  objectiveDefs: LazyDefinition<DestinyObjectiveDefinition>,
-  uninstancedItemObjectives?: {
-    [key: number]: DestinyObjectiveProgress[];
-  }
-): DimObjective[] | null {
-  const objectives =
-    item.itemInstanceId && objectivesMap[item.itemInstanceId]
-      ? objectivesMap[item.itemInstanceId].objectives
-      : uninstancedItemObjectives
-      ? uninstancedItemObjectives[item.itemHash]
-      : [];
-
-  if (!objectives || !objectives.length) {
-    return null;
-  }
-
-  // TODO: we could make a tooltip with the location + activities for each objective (and maybe offer a ghost?)
-  const formatter = formatterSelector(settings.language);
-  return objectives
-    .filter((o) => o.visible && objectiveDefs.get(o.objectiveHash))
-    .map((objective) => {
-      const def = objectiveDefs.get(objective.objectiveHash);
-
-      let complete = false;
-      let booleanValue = false;
-      let display = `${formatter.format(objective.progress || 0)}/${formatter.format(
-        objective.completionValue
-      )}`;
-      let displayStyle: string | null;
-      switch (objective.complete ? def.valueStyle : def.inProgressValueStyle) {
-        case DestinyUnlockValueUIStyle.Integer:
-          display = `${objective.progress || 0}`;
-          displayStyle = 'integer';
-          break;
-        case DestinyUnlockValueUIStyle.Multiplier:
-          display = `${formatter.format((objective.progress || 0) / objective.completionValue)}x`;
-          displayStyle = 'integer';
-          break;
-        case DestinyUnlockValueUIStyle.DateTime: {
-          const date = new Date(0);
-          date.setUTCSeconds(objective.progress || 0);
-          display = `${date.toLocaleDateString()} ${date.toLocaleTimeString()}`;
-          displayStyle = 'integer';
-          break;
-        }
-        case DestinyUnlockValueUIStyle.Checkbox:
-        case DestinyUnlockValueUIStyle.Automatic:
-          displayStyle = null;
-          booleanValue = objective.completionValue === 1;
-          complete = objective.complete;
-          break;
-        default:
-          displayStyle = null;
-          complete = objective.complete;
-      }
-
-      return {
-        displayName:
-          def.displayProperties.name ||
-          def.progressDescription ||
-          (objective.complete ? t('Objectives.Complete') : t('Objectives.Incomplete')),
-        description: def.displayProperties.description,
-        progress: objective.progress || 0,
-        completionValue: objective.completionValue,
-        complete,
-        boolean: booleanValue,
-        displayStyle,
-        display
-      };
-    });
-}
-
-function buildFlavorObjective(
-  item: DestinyItemComponent,
-  objectivesMap: { [key: string]: DestinyItemObjectivesComponent },
-  objectiveDefs: LazyDefinition<DestinyObjectiveDefinition>
-): DimFlavorObjective | null {
-  if (!item.itemInstanceId || !objectivesMap[item.itemInstanceId]) {
-    return null;
-  }
-
-  const flavorObjective = objectivesMap[item.itemInstanceId].flavorObjective;
-  if (!flavorObjective) {
-    return null;
-  }
-
-  // Fancy emblems with multiple trackers are tracked as regular objectives, but the info is duplicated in
-  // flavor objective. If that's the case, skip flavor.
-  const objectives = objectivesMap[item.itemInstanceId].objectives;
-  if (objectives && objectives.some((o) => o.objectiveHash === flavorObjective.objectiveHash)) {
-    return null;
-  }
-
-  const def = objectiveDefs.get(flavorObjective.objectiveHash);
-  return {
-    description: def.progressDescription,
-    icon: def.displayProperties.hasIcon ? def.displayProperties.icon : '',
-    progress:
-      def.valueStyle === 5
-        ? (flavorObjective.progress || 0) / flavorObjective.completionValue
-        : (def.valueStyle === 6 || def.valueStyle === 0 ? flavorObjective.progress : 0) || 0
-  };
 }
 
 function buildTalentGrid(
